// src/types/bulletins/update.tsx
import { data } from './list';
import { BulletinsListStatus } from '../../enums/bulletins/list';

export interface BulletinsUpdatePayload {
  bulletinId: number;
  payload: {
    title: string;
    content: string;
    category_id: number;
    start_date: string;
    end_date: string;
    created_by: number;
    status: number;
    group_id: number;
  };
}

export interface BulletinsUpdateState {
  data: data | null;
  status: BulletinsListStatus;
  error: string | null;
}